+++ packages/browseros-agent/apps/server/src/tools/browser/bookmarks.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
-import type { BookmarkNode } from '../browser/bookmarks'
-import { defineTool } from './framework'
+import type { BookmarkNode } from '../../browser/bookmarks'
+import { defineTool } from '../framework'
 
 const bookmarkNodeSchema = z.object({
   id: z.string(),
