+++ packages/browseros-agent/apps/server/src/tools/browser/tab-groups.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { defineTool } from './framework'
+import { defineTool } from '../framework'
 
 const TAB_GROUP_COLORS = [
   'grey',
