+++ packages/browseros-agent/apps/server/tests/tools/browser-fixtures.test.ts
@@ -4,22 +4,25 @@ import { mkdtempSync, rmSync, writeFileSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { basename, join } from 'node:path'
 import { diffSnapshots } from '../../src/browser/core/snapshot/diff'
-import { search_dom } from '../../src/tools/dom'
+import { search_dom } from '../../src/tools/browser/dom'
 import {
   check,
   click,
   select_option,
   uncheck,
   upload_file,
-} from '../../src/tools/input'
+} from '../../src/tools/browser/input'
 import {
   close_page,
   list_pages,
   navigate_page,
   new_page,
   wait_for,
-} from '../../src/tools/navigation'
-import { evaluate_script, take_snapshot } from '../../src/tools/snapshot'
+} from '../../src/tools/browser/navigation'
+import {
+  evaluate_script,
+  take_snapshot,
+} from '../../src/tools/browser/snapshot'
 import {
   type BrowserFixtureServer,
   fixtureRoutes,
